requirements.txt
black==22.10.0
isort==5.10.1
mypy==0.991
pip==22.3.1
pyproject-autoflake==1.0.2
pyproject-flake8==6.0.0
pytest-asyncio==0.20.2
pytest-cov==4.0.0
setuptools==65.4.1
sniffio==1.3.0
structlog==22.3.0
tox==3.27.1
twine==4.0.2
